+++ src/renderer/opengl/renderer_opengl.hpp
@@ -8,8 +8,8 @@ namespace Vin {
 		void Init() final;
 		void SetViewport(int x, int y, int width, int height) final;
 		void Clear(float r, float g, float b, float a) final;
-		void DrawArrays(const std::shared_ptr<VertexArray>& vertexArray, size_t verticiesCount) final;
-		void DrawIndexed(const std::shared_ptr<VertexArray>& vertexArray, size_t indexCount) final;
+		void DrawArrays(const std::shared_ptr<VertexArray>& vertexArray, uint32_t verticiesCount) final;
+		void DrawIndexed(const std::shared_ptr<VertexArray>& vertexArray, uint32_t indexCount) final;
 		void Blit(const std::shared_ptr<RenderTarget>& src, const std::shared_ptr<RenderTarget>& dst) final;
 		void SetCullMode(CullMode mode) final;
 		void SetBlendMode(BlendMode mode) final;
